+++ falcon/ms_io/mgf_io.py
@@ -9,7 +9,7 @@ import pyteomics.mgf
 import spectrum_utils.spectrum as sus
 
 from ..cluster import cluster
-from . import _reader_utils
+from . import reader_utils
 
 USI_PATTERN = re.compile(r"^mzspec:[^:\s]+:[^:\s]+:(scan:\d+|\d+)(:[^:\s]+)?$")
 
@@ -30,7 +30,7 @@ def get_spectra(source: Union[IO, str]) -> Iterable[sus.MsmsSpectrum]:
         An iterator over the spectra in the given file.
     """
     with pyteomics.mgf.MGF(source) as f_in:
-        base = _reader_utils.base_filename(source, f_in)
+        base = reader_utils.base_filename(source, f_in)
 
         for spectrum_i, spectrum_dict in enumerate(f_in):
             params = spectrum_dict.get("params", {})
@@ -55,7 +55,7 @@ def get_spectra(source: Union[IO, str]) -> Iterable[sus.MsmsSpectrum]:
             try:
                 yield _parse_spectrum(spectrum_dict)
             except (ValueError, KeyError) as e:
-                _reader_utils.log_skipped_spectrum(
+                reader_utils.log_skipped_spectrum(
                     source, params.get("title"), e
                 )
                 continue
