+++ falcon/ms_io/mzml_io.py
@@ -6,7 +6,7 @@ import spectrum_utils.spectrum as sus
 from lxml.etree import LxmlError
 
 from ..config import config
-from . import _reader_utils
+from . import reader_utils
 
 logger = logging.getLogger("falcon")
 
@@ -27,7 +27,7 @@ def get_spectra(source: Union[IO, str]) -> Iterable[sus.MsmsSpectrum]:
         An iterator over the spectra in the given file.
     """
     with pyteomics.mzml.MzML(source) as f_in:
-        filename = _reader_utils.base_filename(source, f_in)
+        filename = reader_utils.base_filename(source, f_in)
         try:
             for spectrum_dict in f_in:
                 if int(spectrum_dict.get("ms level", -1)) > 1:
@@ -46,7 +46,7 @@ def get_spectra(source: Union[IO, str]) -> Iterable[sus.MsmsSpectrum]:
                     try:
                         yield _parse_spectrum(spectrum_dict)
                     except (ValueError, KeyError) as e:
-                        _reader_utils.log_skipped_spectrum(
+                        reader_utils.log_skipped_spectrum(
                             source, spectrum_dict["id"], e
                         )
         except LxmlError as e:
